Authentication status management improved in Header
