+++ src/components/Header/actions.ts
@@ -1,10 +1,8 @@
 "use server";
 
-import { redirect } from "next/navigation";
 import { createClient } from "@/lib/supabase/server";
 
 export async function signOut() {
   const supabase = createClient();
   await supabase.auth.signOut();
-  redirect("/");
 }
